Add capability to add optional suffix to username
Add capability to add optional suffix to username, e.g. @domain.com in the case that you want users to be able to login with just a username and not full email address. While you can customize which field in LDAP contains the username, sometimes you might want to be able to append a predefined suffix.